refactor(cron): tighten shared run admission
