+++ src/cron/service.test-harness.ts
@@ -243,8 +243,7 @@ export function createMockCronStateForJobs(params: {
     activeManualRunJobIds: new Set<string>(),
     manualSetupTimeoutNotified: false,
     runAdmission: { active: 0, waiters: [] },
-    queuedRunReservationAtByJobId: new Map<string, number>(),
-    queuedForceRunReservationAtByJobId: new Map<string, number>(),
+    queuedRunReservationsByJobId: new Map(),
     timer: null,
     storeLoadedAtMs: nowMs,
     op: Promise.resolve(),
